requirements.txt
torch==1.4.0
torchvision==0.5.0
pillow==7.0.0
tqdm==4.42.1
pandas==1.0.1
numpy==1.18.1
opencv-python==4.2.0.32
matplotlib==3.2.0